Slight refactor.
